requirements.txt
asyncpg
fastapi[standard]
uvicorn[standard]
sqlalchemy[asyncio]
alembic
psycopg2-binary
passlib[bcrypt]
bcrypt==3.2.2
itsdangerous
python-dotenv
markdown
jinja2
python-multipart